if block, convert weights,  changes to html
Put the newstory.replace inside the first if block with the name bob, Found the coversions for conversions for centigrade to farenheight and stone to pounds and added that to the second if block. Made some changed to the html with the positioning of the main.js tag but random story generator button still not working.