Several changes
Added the code for darkening and lightening the button. Had problems with getting the images to show up in the thumbnail once they were clicked and realized I had to use newImage in the thumbBar.appendchild and addded thumbar to addevent listener. Also used a simpler method to add src and alt to each image.